backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Data Fetching & Processing
yfinance==0.2.66
pandas>=2.2.3
numpy>=1.26.2
requests==2.31.0
certifi
urllib3>=2.0.0

# Technical Analysis
ta==0.11.0

# File Handling
openpyxl==3.1.2

# Performance & Security (Part 5 Requirements)
pydantic>=2.0.0
Flask-Limiter>=3.5.0
numba>=0.58.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Analysis & Refactoring (Part 5C)
graphviz>=0.20.0
